requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.34.0
supabase==2.10.0
pyjwt[crypto]==2.10.1
pydantic==2.10.4
python-dotenv==1.0.1
